+++ packages/ice-npm-utils/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.2.0
+
+- [feat] getNpmRegistry 优先读取 npm config，否则返回 `https://registry.npm.com`
+- [fix] extract tarball consider directory
+
 ## 1.1.2
 
 - [fix] 修复getAndExtractTarball写空文件时卡死的问题